made login the entry point
